+++ src/flint.io/flint/machine/AnalogInput.java
@@ -2,7 +2,7 @@ package flint.machine;
 
 import java.io.IOException;
 
-public interface AnalogInputStream {
+public interface AnalogInput {
     int read(byte[] b) throws IOException;
     int read(byte[] b, int off, int count) throws IOException;