+++ src/flint.io/flint/machine/OneWire.java
@@ -1,6 +1,8 @@
 package flint.machine;
 
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 public class OneWire implements InputPort, OutputPort {
     private String name;
@@ -75,4 +77,12 @@ public class OneWire implements InputPort, OutputPort {
 
     @Override
     public native void write(byte[] b, int off, int count) throws IOException;
+
+    public InputStream getInputStream() {
+        return new InputPortStream(this);
+    }
+
+    public OutputStream getOutputStream() {
+        return new OutputPortStream(this);
+    }
 }